# docs/swagger-doc.yaml
openapi: 3.0.1
info:
  title: Tunemyles API
  description: >-
    This is a sample server Petstore server. You can find out more about Swagger
    at [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net](irc.freenode.net)
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'http://localhost:5000/api/v2'
    description: Dev Server API URL
...
paths:
...
  /account/merchant:
    post:
      tags:
        - account
      summary: Activate User Merchant (Route All Roles)
      description: Activate User Merchant (Route All Roles)
      operationId: activateMerchant
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ActivateMerchantBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuthorizedResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
  /merchants:
    get:
      tags:
        - merchants
      summary: Get Single Merchant (Merchant Route)
      description: Get Single Merchant (Merchant Route)
      operationId: getSingleMerchant
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMerchantProfileResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
    patch:
      tags:
        - merchants
      summary: Update Merchant Data (Merchant Route)
      description: Upadate Merchant Data (Merchant Route)
      operationId: updateMerchant
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateMerchantProfileBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  /merchants/operation-time:
    patch:
      tags:
        - merchants
      summary: Update Merchant Operation Time (Merchant Route)
      description: Update Merchant Operation Time (Merchant Route)
      operationId: updateMerchantOperation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMerchantClosetimeBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
  /merchants/products:
    post:
      tags:
        - merchants
      summary: Add Product (Merchant Route)
      description: Add Product (Merchant Route)
      operationId: addProduct
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PostMerchantProductBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostMerchantProductResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
    get:
      tags:
        - merchants
      summary: Product List (Merchant Route)
      description: Product List (Merchant Route)
      operationId: productList
      parameters:
        - name: orderDirection
          in: query
          schema:
            $ref: '#/components/schemas/OrderDirectionQuery'
        - name: orderBy
          in: query
          schema:
            allOf:
              - $ref: '#/components/schemas/OrderByQuery'
              - enum:
                  - created_at
                  - product_name
                  - market_name
                  - market_address
        - name: search
          in: query
          description: 'Available search field product_name, market_name, market_address'
          schema:
            $ref: '#/components/schemas/SearchQuery'
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/PageQuery'
        - name: limit
          in: query
          schema:
            $ref: '#/components/schemas/LimitQuery'
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMerchantProductsResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  '/merchants/products/{id}':
    patch:
      tags:
        - merchants
      summary: Update Product (Merchant Route)
      description: Update Product (Merchant Route)
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Request Data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMerchantProductBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
    get:
      tags:
        - merchants
      summary: Get Single Product (Route All Roles)
      description: Get Single Product (Route All Roles)
      operationId: singleProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMerchantProductDetailResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
    delete:
      tags:
        - merchants
      summary: Delete Product (Merchant Route)
      description: Delete Product (Merchant Route)
      operationId: deleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  '/merchants/products/{id}/cover':
    patch:
      tags:
        - merchants
      summary: Change Product Cover (Merchant Route)
      description: Change Product Cover (Merchant Route)
      operationId: productCover
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateMerchantProductCoverBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeMerchantProductCoverResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  '/merchants/products/{id}/status':
    patch:
      tags:
        - merchants
      summary: Change Product Status (Merchant Route)
      description: Change Product Status (Merchant Route)
      operationId: productStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMerchantProductStatusBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateMerchantProductStatusResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
  '/merchants/products/{id}/category':
    patch:
      tags:
        - merchants
      summary: Bind Product Category (Merchant Route)
      description: Bind Product Category (Merchant Route)
      operationId: bindProductCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BindMerchantProductCategoryBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BindMerchantProductCategoryResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
      x-codegen-request-body-name: body
  '/merchants/products/{id}/image':
    post:
      tags:
        - merchants
      summary: Add Product Image (Merchant Route)
      description: Add Product Image (Merchant Route)
      operationId: productImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PostMerchantProductImageBody'
        required: true
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostMerchantProductImageResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
  '/merchants/products/image/{id}':
    delete:
      tags:
        - merchants
      summary: Delete Product Image (Merchant Route)
      description: Delete Product Image (Merchant Route)
      operationId: delteProductImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
  /merchants/orders:
    get:
      tags:
        - merchants
      summary: Merchant Orders (Merchant Route)
      description: Merchant Orders (Merchant Route)
      operationId: merchantOrders
      parameters:
        - name: orderDirection
          in: query
          schema:
            $ref: '#/components/schemas/OrderDirectionQuery'
        - name: orderBy
          in: query
          schema:
            allOf:
              - $ref: '#/components/schemas/OrderByQuery'
              - enum:
                  - created_at
                  - full_name
                  - phone_number
                  - address
        - name: search
          in: query
          description: 'Available search field full_name, phone_number, address'
          schema:
            $ref: '#/components/schemas/SearchQuery'
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/PageQuery'
        - name: limit
          in: query
          schema:
            $ref: '#/components/schemas/LimitQuery'
        - name: status
          in: query
          schema:
            allOf:
              - $ref: '#/components/schemas/StatusQuery'
              - enum:
                  - '0'
                  - '1'
                  - '2'
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMerchantOrdersResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - APIKeyHeader: []
...
  /merchants/list:
    get:
      tags:
        - merchants
      summary: Merchant List
      description: Merchant List
      operationId: merchantList
      parameters:
        - name: orderDirection
          in: query
          schema:
            $ref: '#/components/schemas/OrderDirectionQuery'
        - name: orderBy
          in: query
          schema:
            allOf:
              - $ref: '#/components/schemas/OrderByQuery'
              - enum:
                  - created_at
                  - full_name
                  - phone_number
                  - market_name
                  - market_address
        - name: search
          in: query
          description: >-
            Available search field full_name, phone_number, market_name,
            market_address
          schema:
            $ref: '#/components/schemas/SearchQuery'
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/PageQuery'
        - name: limit
          in: query
          schema:
            $ref: '#/components/schemas/LimitQuery'
      responses:
        '200':
          description: Request Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMerchantListResponse'
        4XX:
          description: Request Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        5XX:
          description: Sever Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
...
    OrderDirectionQuery:
      type: string
      default: DESC
      enum:
        - ASC
        - DESC
...
    ApiResponse:
      type: object
      properties:
        code:
          type: number
        success:
          type: boolean
        messages:
          type: string
...
    UserAuthorizedResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                type:
                  type: integer
                token:
                  type: string
...
    ActivateMerchantBody:
      required:
        - identity_photo
        - market_photo
        - no_identity
        - market_name
        - market_address
        - market_lat
        - market_lon
        - market_close_time
      properties:
        identity_photo:
          type: string
          format: binary
        market_photo:
          type: string
          format: binary
        no_identity:
          type: string
        market_name:
          type: string
        market_address:
          type: string
        market_lat:
          type: number
        market_lon:
          type: number
        market_close_time:
          type: string
...
    GetMerchantProfileResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - $ref: '#/components/schemas/MerchantResponse'
...
    UpdateMerchantClosetimeBody:
      type: object
      properties:
        close_time:
          type: string
...
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      name: authorization
      in: header